Add loader_json.h/.c for custom loader code
Several functions have been added to the cJSON.h/.c files. While these
functions are important, they do not belong in the cJSON source code
files, as it makes it difficult to update cJSON without stomping all
over the added code.

This does require re-adding cJSON_Parse to the cJSON.h header file,
which was previously removed as it was only used internally to
cJSON.c. Now that it isn't the case, it is needed again.
